+++ include/smokerand/hwtests.h
@@ -3,7 +3,7 @@
  * @brief Hamming weights bases tests implementation, mainly based
  * on overlapping tuples and similar to DC6 from PractRand.
  * @copyright
- * (c) 2024-2025 Alexey L. Voskov, Lomonosov Moscow State University.
+ * (c) 2024-2026 Alexey L. Voskov, Lomonosov Moscow State University.
  * alvoskov@gmail.com
  *
  * This software is licensed under the MIT license.